[webui] Fix 'undefined method can_modify_project? for nil:NilClass' in BsRequestActionWebuiInfosJob
The issue was in the DelayedJob is no User.current set which causes that BsRequestAction.source_access_check!
in SourcePackageFinder to throw a 'undefined method can_modify_project? for nil:NilClass'
for packages coming from a remote or linked project as they can not be found with Package#find_by_name_and_project.
This Job is only to warm up the caches in the backend by sending the diff requests and immediately discard it.
Therefore we can skip the access check anyway as the result won't be displayed somewhere.

Fix #4404