[frontend] Fix encoding of sourcediffs
When the frontend requests a sourcediff that is not in xml format the
returned string is not encoded. Which means that the content can have
differently encoded characters, eg. a tarball that contains diffferently
encoded files.
When handling such a sourcediff string in OBS this could cause encoding
issues. For example sourcediff.present? would raise an
ArgumentError: invalid byte sequence in UTF-8.

To solve this we replace non-UTF-8 encoded bytes with '?', the
default replacement string defined by String#encode.

https://ruby-doc.org/core-2.2.0/String.html#method-i-encode

A note for the VCR configuration change:

Since the backend doesn't return properly encoded data, VCR would fail
on this when handling the backend response. To prevent this we have to tell
VCR to not encode such requests as described here:
  http://www.rubydoc.info/github/vcr/vcr/VCR%2FConfiguration%3Apreserve_exact_body_bytes