[frontend] Fix encoding issue in BsRequestAction#contains_change?
The contains_change? method was comparing the full diff of source and
target package. This failed when the diff contained files with different
encoding.

Another problem of this was that fetching the full source diff could be
expensive depending on the size of the sources.

To solve this we fetch the source info of source and target package. If
both hashsums are identical, we know the sources didn't change.