[frontend] Add .bundle & vendor/bundle directories to .gitignore
in .bundle directory is the local bundle configuration stored,
e.g. the number of retries we change in test so we do not
want to check this into git.
vendor/bundle stores all the gems downloaded when running the tests
we also don't want to check this into git.
Both directories are in the default .gitignore for Rails:
https://github.com/github/gitignore/blob/master/Rails.gitignore