[ci] Fix Rubocop offense Metrics/LineLength
When generating the `.rubocop_todo.yml` file, it outputs the following
warning:

Phase 1 of 2: run Metrics/LineLength cop (skipped because the default
Metrics/LineLength:Max is overridden)

It is so because we define `Max: 150` under `Metrics/LineLength` in
`.rubocop.yml`. In the end, these offenses are not put in
`.rubocop_todo.yml`. This is why I fixed it