[frontend] Fix test for forwarding bs requests
This test was failing with a MissingAction exception when the VCR cassettes
where not used. To fix this we had to ensure that the target package of
the forwarded request has different source content than the original
package.

In our test suite this test was still passing because of the recorded
cassettes.