[frontend] Include request action validations to bs request validations
Before it wasn't possible to see why a bs request action was causing a
bs request to fail. All a user would see was 'Validation failed: Bs
request actions is invalid'.
Now the validation message would include details of the failed request
action.

This allows users to see why a request failed, eg. when forwarding a
request (related to #3730).