[webui] Fix multibuild binary view
We were using the `Package` object for the multibuild view, but the
multibuild is just a name, that's why the cointainer was shown.

Now the multibuild package name is given to the backend call to get the
correct results.

I also updated the links in the view, except by the `Job History`, as this
view only exist for the container at the moment.