[frontend] Add a way to switch to the new WebUI for certain actions
A way of showing the new WebUI only for some users (admins and staff)
was needed. We provide a method called 'switch_to_webui2' that does
the switch of layouts and views needed.

At the same time this method calls the "ovewriting" method (properly
named with a prefix 'webui2_'), something like webui2_NAME_OF_ACTION.