[backend] Add new regex needssslcertforbuild for Debian builds
The only programmatic supported way to add arbitrary strings to a Debian
dsc file is via X[SBC]-Key: value tags in debian/control.
The regex that checks if the SSL certificate needs to be copied into the
build root checks that there is exactly "# needssslcertforbuild", with
the hash sign having to be the first character of the line.
But in Debian at best we can do:
  Obs: needssslcertforbuild
by adding:
  XS-Obs: needssslcertforbuild
to debian/control.
So add a new regex for this tag. This will be extensively documented in
pesign-obs-integration.

(cherry picked from commit 4442b9f5a3bf0e8aaf6bb2f3215c0062391eeccf)