[frontend] Document how to handle delayed jobs queue during updates
When updating OBS, it can happen that the delayed jobs table contains
pending jobs. In case that the OBS update includes a code change that
is incompatible with the object stored by delayed jobs, this can cause
breakage when once the delayed job gets executed.

This can either happen when running database migrations (see PR#4969)
or by processing the delayed jobs service.

While this can happen for any deployed OBS instance it is far more
likely this occurs in our Stable releases. This is because of the much
larger delay between creating a migration and running it (= updating the
OBS instance).

Since this can hardly be avoided, we document this limitation and how to
avoid this from happening in the updaters guide.

Fixes #4969