[frontend] Explicitly set charset encoding for tables
Split out of https://github.com/openSUSE/open-build-service/pull/4809

Our db migration test is comparing the structure.sql created with the
structure.sql we store in our git repository. This happens for any of our
rpm package builds. Which means that this test has to pass with various
MySQL versions (we don't use a schema.rb :cry:).
Because of that we either have to explicitly set defaults or workaround
the migration test to make it pass.

This commit adds a migration to explicitly set the charset encoding for
tables and columns that were previously using the default settings.
This means that the OBS structure.sql will now look the same on different
systems (aka with different MySQL configurations), eg. SLE15 vs SLE12.
And therefor pass our structure sql test.