[api][ci] Normalize default values for text types
From MariaDB 10.2.1 on TEXT and BLOB data types can have a DEFAULT value.
As a result, from MariaDB 10.2.1, SHOW CREATE TABLE will append a DEFAULT NULL to nullable TEXT or BLOB
fields if no specific default is provided.
Therefore We normalize the structure.sql dump for the structure verify test as
long as we support SLE12/Leap 42.3 and SLE15.
https://mariadb.com/kb/en/library/show-create-table/