[api][ci] Add collate to all tables
with SLE15 the character_set_server and collation_server changed
from utf8 to utf8mb4.
Tables which do not have a character set / collation defined
will use the server character set / collation.

This caused that when updating from SLE12 to SLE15,
all tables which do not had a collation changed
the character set for the table to be compatible
(utf8 character set does not work with utf8mb4_bin collation).

https://mariadb.com/kb/en/library/setting-character-sets-and-collations/