[dist] Use multiple arguments in our Rakefile
This prevents us from interpolating user input.

Our rakefile is passing certain environment variables to the
docker-compose command. Since we were not escaping these parameters
it was possible to manipulate or break the command, eg. by having
a ' character in a commit message.

Found due to a failure in travis:
  rake docker:test:lint
  ...
  sh: 7: Syntax error: Unterminated quoted string