[ci] Fix travis failure script
In case of test failures our travis failure script switches to the
log directory and uploads all files it finds there.
With the switch to docker the start directory travis is running from
changed and changing to the log dir failed silently.

Because of that we uploaded all files within our git repo and not only
our frontend log files.

This commit updates the cd command and causes the script to fail if the
log directory can not be found.