[dist] Silence hidden-file-or-dir rpmlint check
Some of our files are intenionally hidden, eg. the AWS configurations,
the sprockets manifest, and many of files of the gems we bundle.
Because of that we have a lot of false positives so that we would
probably overlook offenses we might want to fix.
And given that this offense doesn't cause any harm it's better to
silence this check.
That makes other, more critical offenses more visible.