[dist] Rpmlint: Silence files-duplicate warnings
Since we soon switch to the bundle gems service we need to solve
some rpmlint issues caused by the new gem files we store in the
vendor/ directory.

Many of the bundled gems cause a files-duplicate warning. Since we
don't want to mess around with the gem sources we silence the check.