[dist] Rpmlint: Silence script-without-shebang warnings
Since we soon switch to the bundle gems service we need to solve
some rpmlint issues caused by the new gem files we store in the
vendor/ directory.

Many of the bundled gems we package contain files with executable
bit. Since we won't be able to solve all them upstream in reasonable
time we disable that check.