[dist] Rpmlint: Silence unexpanded-macro warnings
Since we soon switch to the bundle gems service we need to solve
some rpmlint issues caused by the new gem files we store in the
vendor/ directory.

The rails gems have a couple of '%namespaced_name%' tags in their
filenames. There is nothing we can do about this. So we silence
this check to not spam the rpmlint log.