[ci] improve debugging output in 0030_create_and_build_package.ts
With this commit the output of 0030_create_and_build_package.ts in case of
an timeout should be improved, so you can see that a timeout occured and the
last building state of the project at a glance.

The variable OBS_TEST_TIMEOUT is required to reduce the timeout while
testing the script itself.

Esp. the mails/web output of kanku should be more meanful with this patch.