[frontend] Introduces switch to a new layout
For certain actions and certain users we want to switch to a new bootstrap
based layout while for everyone else things should stay the same.

Adding the `switch_to_webui2` method to a controllers `index` action will

1. switch the layout to bootstrap
2. switch the view path to files below `app/views/webui2`
3. call the `webui2_index` method of the controller (if it exists)

That means you are flexible to overwrite only how the action is rendered (by
adding `switch_to_webui2` to the end of your controller method) or can also
overwrite what you do in the action (by putting it on the top of your method).

Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>
Co-authored-by: Moises Deniz Aleman <mdeniz@suse.com>
Co-authored-by: Stasiek Michalski <hellcp@opensuse.org>