[ci] Implement feature test suite for bootstrap
During the period of migrating our views to bootstrap, we need to run the feature tests twice.
One time with bootstrap disabled and one time enabled.
This can be triggered by setting the BOOTSTRAP environment variable and
excluding the spec/bootstrap/ directory from rspec (see .circleci/config.yml).
If a test does not run with reasonable effort with bootstrap enabled,
the test can be skipped by adding skip_if_bootstrap to the first line of the spec
and we can implementing a new test in spec/bootstrap/ directory.