[ci] Split up rubocop configurations
Previously we were running rubocop once from within the rails directory,
but linted all files in our git repository. To make that possible we had
a main rubocop config in the top directory and a todo file in the top
and rails dir.
This caused a couple of problems, eg. excluded files would have incorrect
relative path. This setup also caused us to configure rails cops
globally (aka for non-rails files).

Splitting the configuration up makes the setup simpler and better to
maintain.

Co-authored-by: Eduardo Navarro <enavarro@suse.com>