[dist] Change default worker run directory
Current default run directory for OBS workers is under /var/run. While
in most modern distros Systemd is the responsible for mounting /run and
does so with executable permissions, Debian (and Debian downstream)
distributions delegate this task to their initramfs package, which
mounts /run with the noexec flag on. Note that /var/run is a symlink to
/run on such distributions. This commit changes the default run
directory to a path under /tmp to solve the described issue while
keeping the volatile properties of the directory under /run. See
https://freedesktop.org/wiki/Software/systemd/APIFileSystems and
https://www.freedesktop.org/wiki/Software/systemd/InitrdInterface/ for
further reference.