[frontend] Simplify storage of attributes into backend
We need the _attributes file in the backend for revision
storage, but the only code that actually wrote was the
attribute controller in the api. All attributes created,
updated or deleted through the webui weren't recorded
as revision.

So simplify the storage and let the model handle it. This
way we don't have to duplicate the code too much.

I added a write_attributes to the webui controller though
as the API has tests that the revision is not increased
on commits that do not update values (which is a bit questionable,
but I didn't want to break tests)