[frontend] Drop BuildReasonFile model
The BuildReasonFile model was abstracting the _reason file that is
stored in the backend.
Though the only place where this method was used we parsed the raw xml
and created another object (PackageBuildReason) out of it.

Because of that it makes sense to drop this model and use a Backend::Api
call like we already do elsewhere.

In addition this solves an issue caused by BuildReasonFile#to_s returning
nil when the file isn't available or invalid.

Kudos for the idea goes to @DavidKang