[api] Do not use succeed + surround
`succeed` appends a string to the end of a Haml block, and `surround`
appends strings to the end and the beginning. Because of that, if we
already need to use `surround`, it doesn't make sense to use `succeed`
as well. It is also more readable, as what we want is just a comma after
the closing parenthesis.