[api] Use superseding.last instead of index
We want a comma except for the last record, because of that I think that
it is more readable to use `superseding.last` than using an index and
comparing it with `superseding.length - 1`. This way we also don't need
to use `each_with_index`, as we don't really care of what index every
record have.

This may be a matter of taste, but I find it more elegant.