[ci] Use in memory database in CI
In Travis we used a tmpfs volume before.
Using tmpfs volume seems not to be possible in CircleCI
as not all native docker / docker-compose commands are supported.
The official CircleCI database images mount as well /dev/shm
to make use of an in memory database.

Changing Travis is mostly done to maintain the same setup in
CircleCI and Travis.

We also need to call the command directly in the config.yml
and the database container does not have a git checkout,
therefore calling ./contrib/start_test_db is not possible.

See for the official CircleCI images:
https://hub.docker.com/r/circleci/mysql/tags/