[ci] Enable rspec-retry again when explicit set for a feature test
It is now possible to flag a test with retry: 3 to also retry in CI.
Before it was only enabled for package builds in OBS.
However, we had some flaky tests in CI which we now explicit set.
This makes it possible to maintain a list of flaky tests with the advantage
of running these flaky tests nevertheless.