[backend] bs_dodup: support deb query strings in URLs
The debian distro and components are currently parsed by taking trailing
parts of the URL path split with /. This means that the distro and
components cannot contain /s, but both are legal characters in debian
repository specifiers. For instance, the debian security repository uses
distro names of <release>/updates as described in
https://www.debian.org/security/#keeping-secure.

To support this, allow the URL to specify the distro and components in a
query string. For example, a URL for the above security repo would be
http://security.debian.org/debian-security?distro=stable/updates&components=main.

https://github.com/openSUSE/open-build-service/issues/5217