[frontend] Fix rubocop offense: Naming/PredicateName
After updating the initial implementation of RuboCop rake tasks for OBS (#5173)
we switched to using the RuboCop::RakeTask library (#5182).
This change caused that the *:all task was only running one of the two
RuboCop setups in our test suite.
PR#5219 solved that issue and this commit cleans up the remaining
RuboCop offenses.