[frontend] Add User#can_modify? method
We have a couple of code pathes in OBS that operate on both,
projects and packages. So we end up many times having to check whether
we have to run a permission check on an a project or package.

This commit adds a method that calls can_modify_package? or
can_modify_project? depending on the given object class.