[ci] Regenerate rubocop_todo.yml
To exclude DateTime offenses. I actually wanted to disable it in the
`.rubocop.yml` file, because although we could use Time to make Rubocop
happy, as both classes can be used to do the same, the `to_s` is
different. So it seems reasonable to keep it for compatibility reason.
But is is not possible because of a Rubocop bug:

https://github.com/rubocop-hq/rubocop/issues/5199