[webui][api] Fix delete request diff for packages
The if statements were completely messed up.
This got introduced by a refactoring in
214ba835ddcb6fc02a216dd3108c27c1ad12f724.

The DiffError should only raise if there is no
package and no repository (messed with the unless).
Also the guard clause should only return and empty
string if there is no target_package.
In case there is package, we can do a normal diff.

Fix #5107