[ci] Set frontend user in CircleCI
because we changed to use Docker to build the images in OBS.
Therefore we set now the root user for the frontend-base image
to be similar to the development frontend-base image.
Therefore we need to switch now to the frontend user in CircleCI
to execute the tests not as root user.