[webui] Run puma on port 3000 in dev
After we switched from unicorn to puma in #5284, the rails development server
started to run on port 5000.
This is caused by foreman as it sets the  environment variable to 5000.
There is also a bug in puma which does not respect the puma.rb configuration
when the  environment variable is set.
https://github.com/puma/puma/issues/939

Furthermore we need to bind now to 0.0.0.0 as puma binds by default to localhost
which is not exposed by docker.