[webui] Use #prop instead of #attr
As of jQuery 1.6+ #prop should be used to set the property of boolean attributes.
The checked attribute does not correspond to the checked property.
The attribute actually corresponds to the defaultChecked property and should be used only
to set the initial value of the checkbox.
The checked attribute value does not change with the state of the checkbox,
while the checked property does.
Same applies for disabled as well.

See "Attributes vs. Properties" at https://api.jquery.com/prop/ for full reference.

Fix #5294