[webui] Show all open reviews at top of request page
and remove the links from the decision comment tabs.

Because this 'broke' the tabs as half of the tab was a link
to the user/group/package/project show page and
the other half was the actual tab change.
For user reviews it also included the user icon which looked broke the CSS.
The links are now at the top of the page so no need to add it to the tab.

Fix #4679