[ci] Do not report indivual coverage reports but merge manually
This allows several benefits:
 - we always have HTML reports within circleci artefacts
 - we avoid bugs in merges on coverage services
 - the coverage services don't have to wait - if we upload,
   we're done
 - the coverage can be grouped as wildly as we need it,
   it can be done within the merge task

Codecov still does not report status back reliable, but coveralls
seems to have no problem with it. It doesn't hurt to report into
both though - if we let coveralls not comment but only report
status we can have codecov comments as optional. It will be 100%
the same coverage we calculated ourselves on circleci

Reused the existing lib/tasks/coverage.rake that wasn't in used
since 2012 (when travis+coveralls became a valid option)