[dist] Fix definition of search indexes
The definition of the search indexes worked partially before, only for
the 'name' fields.

As it is said in the documentation the definition of each index follows
this pattern:

indexes field_name

Applying this patch makes fields 'title' and 'description' be indexed
because now the definition of each index is followed, as described in the
documentation.