[frontend] Fix undefined method error in user helper
This method got recently moved and refactored (15e02640b95cf), which
caused a bug in the _printed_name method.

"undefined method `<<' for #<User:0x00007f9f368b6dc0>"

We have to duplicate the user login (`user.login.dup`), because
otherwise we would assign the object reference of user.login. Which
means that any later changes of `printed_name` would alter the login of
the user object, e.g. (when we call String#<< later in this method).