Move request creation from request controller to model
This removes quite some duplication in the controller.
The conditional setting of optional request action attributes could be
removed because the default values of these attributes are null.
So setting source_package to params[:devel_package], where params[:devel_package]
is nil, won't alter the created object.

The explicit setting of request state 'new' got dropped, because there
is a callback that sets any newly created request to state new.