[ci] Don't reseed the database on every test start
As followup on #5327 - reseeding the database on every
test run is super conservative and steals a lot of developer
time for the sake of being safe on very rare test crashes

Truncating the static tables is good enough for normal test
crashes. The model tests run in transactions, so don't
interfere with the database - but are the ones most likely
to break.

And in case the database is busted, recreating it is just
as fast/slow as what we basically did before every test
run.

The reseeding takes on my system ~10 seconds - and according
to dmarcoux's comment in 5327 on his it's a minute. So drop it