[db] Hardcode encoding in old migrations
The old migrations are tested on current database and database
settings. This makes it bound to the current encoding - and that is
utf8mb4. But without Barracuda tables you can't have varchars. So hardcode
the encoding used in 2017 by using execute instead of rails helpers

On create it's easy to add, but renames and other column changes
need to be changed to executes as they otherwise drop the charset