Render views in staging projects controller tests
Rendering the views allows us to test the full functionality of the
staging board.

This would have revealed the regression fixed in #5446.

Some of the views require data from projects, requests and other
objects that rely on the backend, this also means we have to use the
global write through and record cassettes.