Render views in distribution controller tests
Rendering the views allows us to test the full functionality of the
staging board.

The Tumbleweed distribution has two rings, Bootstrap and MinimalX.
Enabling the view rendering required to create a MinimalX project
to avoid a Nil error when the RINGs are shown in the view.

The fake OpenqaJob was causing an error as well. Instead we don't return
any OpenqaJobs for now.

```
  ActionView::Template::Error:
    undefined method `result_or_state' for :fake:Symbol
  # ./app/views/webui/obs_factory/distributions/show.html.haml:91:in `each'
```

Some of the views require data from projects, requests and other
objects that rely on the backend, this also means we have to use the
global write through and record cassettes.