Fix Security/Open Rubocop offense
Using openURI's open method allows access to Kernel#open, which is
a bad idea when the parsed URIs can be provided by users.
This isn't the case for us right now, but since we plan to make the
staging dashboards configurable it makes sense to solve this now.
Otherwise this might become a problem in the future.

https://sakurity.com/blog/2015/02/28/openuri.html