Fix StatusControllerSpec to work on empty package
The test is using backend data - and if sticking
to a fixed package name we use in other tests too,
we can't guarantee it's working on empty sources.

And as soon as there are sources, you have a verifymd5

The alternative is to delete the package and recreate
it again, but that appeared overkill. Note that this
fix only works without cassettes