[backend] prevent bs_regpush from crashing if no tags in registry
This patch prevents bs_regpush from crashing if no tags are available
in the registry.

The registry could return a json like the following:

{"name":"home/m0ses/mycontainer","tags":null}

which results in an error message like

Can't use an undefined value as an ARRAY reference at /usr/lib/obs/server/bs_regpush line 233.

and bs_regpush die's

With this patch we set an empty ARRAY reference if $r->{tags} is empty.