Introduce Status::ChecksController#create
action because there is no reason to create and update in the same action.
Also it did not work like it was before as it always created a new Check and
never updated existing ones.
Also moved parsing of the XML body to a before filter which renders an error
in case it is empty.