Add Terms of Service to the views
Terms of Services are shown on any page of the Webui until a user
acknowledges them.
When there are new ToS the ToS notification message appears again and
the user has to acknowledge the new version.
The show page provides the full text of the ToS, while the notification
message just contains a link to it.

Additionally:
* Terms of Services / announcement can be rendered as markdown.
* The order of ToS versions is tracked via it's creation time.
* The footer contains a 'Terms of Services' that links to the most
  recent ToS, if there are any.

Co-authored-by: Björn Geuken <bgeuken@suse.de>
Co-authored-by: David Kang <dkang@suse.com>