Use cache keys to better determine changes of cached objects
The access check of the public controller relies on some cached data.
Since the used cache key was only using the project and package names
the cached data could easily be outdated by changes of the objects, eg.
issue #5418 can be caused by using the cached data even though the
package got deleted meanwhile.

The caching was added in fe523c18082b844fd2ab720636962c2e021086a8
because of heavy usage of this endpoint and we still gone need this.
Therefore we extend the code to use the package and project cache keys.