Add run_rpihybridiso flag to product
We're growing support to enable our installation .iso to boot directly
on the Raspberry Pi. For that we need to craft a special MBR which we
need to trickle down from the product description into the kiwi xml.

This patch adds awareness for a new product definition flag to indicate
that we want to build an iso image that can boot on the Raspberry Pi.

Signed-off-by: Alexander Graf <agraf@suse.de>