Implement Status::RequiredChecks
A required check can be defined on a checkable
(e.g. project, package, repository). When a check
does not exist, the required check will be displayed instead.
As soon as a check exists with the same name as the required
check, this check will be displayed instead.

For now, only repository as checkable is supported and the
required checks are represented as a serialized array on
repository.

In the WebUI this is implemented on the staging project dashboard.

Co-authored-by: Dany Marcoux <dmarcoux@suse.com>