Fix missing attribute error in statistics controller
In 429c1b2 we include some additional data to the json representation of
bs requests. This was needed by the staging controller / openSUSE release tools.
Though when the superseded_by attribute is not defined, eg. by fetching
a subset of attributes from the database, this would cause a 'missing
attribute' error.

Checking whether superseded_by is defined before calling it fixes the
issue.

Fixes #5761