Introduce a status report model
The status report model contains a collection of status checks and
belongs to a polymorphic checkable object.

The uuid / uuid_type field provide a field for storing entities that
are not represented by a model or other class in our frontend, eg. the
build id of a published repository (stored in the backend).

The status report routes are wrapped by a concern and can be attached to
various objects. For now we support repositories and bs_requests to have
status reports.

In addition this:
* Adds tests for status related policies
* Extends and refactor tests status check controller
* Replaces RepositoryPublish model with Status::Report