Fix 500 error on attribute edit page
When an attribute got deleted from a project or package, or did not
exist in first place, the @attribute variable would unexpectedly be nil.
With this commit we raise an ActiveRecord::RecordNotFound in such a
case.
The exception will be catched by an exception handler which renders
a generic 404 page, in production mode.