Refactor render_dialog for change_devel_request_dialog
Instead of performing an Ajax call that answers with a Javascript code
that updates the view to create and show the modal we now render the
modal if needed and show it via Javascript locally.

This is better because we only need one SQL query more and we save a XHR
request.

Co-authored-by: David Kang <dkang@suse.com>