Fix tests for required checks controller
The tested http requests was only added to the first it-block of
some of the shared examples.
Because of that the following it-blocks were testing the login requests
(test preparation) instead of the actual controller action.

This commit fixes it by moving the tested http request into a subject
and re-using it in every it-block.