Migrate Package#repositories page to bootstrap
Instead of showing complex icons we have simplified those to only show
the actual status of it (either enabled or disabled).

We made this page responsive to able small devices user's to work here.

Now the spinner is shown in place (replacing the icon of the flag)
until it finish.

The actions that can be performed per flag are now inside a popover
(needs to click on the icon flag instead of just hovering). This popover
shows in the title the actual status of the flag (and if this is a
default value or not).

Co-authored-by: Björn Geuken <bgeuken@suse.de>
Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>
Co-authored-by: Lukas Krause <lkrause@suse.de>