Add a buildable package to our review app setup
We are using review apps for reviewing UI changes on a temporary
spawn OBS instance. These instances are filled with data based on
the dev:development_testdata rake task.
This commit adds some buildable source to the package, which ensures
that any spawn review app has some built binaries that we can use for
testing.