Remove stubs from test for BsRequest.delayed_auto_accept
The test was stubbing the auto_accept call and veryfied the method call
afterwards. That is not a problem in itself, but doing so required a
couple of additional stubs to pass along the 'request' we were testing.
Because of that the test was relying to much on internals of the tested
code.
Removing the stubs makes the test more flexible of changes in the
auto accept functionality and ensures that the full auto accept workflow
is tested.