Fix package add file validations
The JS validation did not take into account when there was already
another input field with a value. Before it just set the required
tag on all input fields when the value was empty. This was not correct
as it could happen that there is already another input field with a
value. In this case we don't need to change the required tag at all
and just leave it like it is.

Fix #5980.