Fix colors for links that emulate outline buttons
Adding to an anchor/link outline button style, the texts are shown in
black. The color of the text inside the cancel link/button should be the
same as its boder's color or white when hover. Adding the class
`btn-outline-*` to the element is not enough, so we make them extend
from `btn-outline-*` in CSS.

Co-authored-by: David Kang <dkang@suse.com>