Do not show 'View file' link for special files
before we used a unless_link_to helper, this was c&p from the old view.
However, the unless_link_to helper returns only the link text in
case it evalutes to false. In this case we don't want to show the link
at all anymore therefore we removed the unless_to_link helper and
check for the condition before we even render the link.

#5987.