Fix burger menu and show breadcrumbs in mobile
Before we had different watchlist views depending on the size of the
device, for mobile was broken. Now we only use one single watchlist
component.

The breadcrumbs was missing in mobile now is shown normally.

The personal navigation was shown in the burger menu badly, now is shown
in the same way as in big screen devices.