Hide user icon for XS displays
To ident the comments we use bootstraps media object.
However, on small screens the indentation does not make sense
as the input field gets smaller and smaller. Therefore we just
disable the indentation on XS screens. This is done by applying the
bootstrap display CSS classes to the user image.

Fix #5981.