Limit filename to fix button overlapping on revision diff page
On small screens the buttons overlapped the filename which
gets fixed by limiting the filename to 75% width.

Additionally the buttons gets now disabled and not hidden
when it is a binary etc. This improves usability as the mouse
is not over a different button when using the jumping buttons
next to it.

Fix #5987