[backend] replace haveinterrepodep code with something more useful
We used to just check if there was a dependant repository in the
same project. This works for some use cases like maintenance
incidents, but misses a lot of other cases like the openSUSE
ports setup.

We now also check projects that are "related", i.e. they are either
upstream or downstream in the project hierarchy list.