Suppress additional rpm output on package update
We recently changed to use systemd instead of SysV for starting delayed
jobs services. Part of this was adding a script that ensures that the new systemd
service would be added when updating packages (710cd0fe48bb4a63212dee17a3fcb8b2bcdf6cbe).

This was working fine for the initial switch from SysV to systemd, but
once the SysV service is removed any following package update would
cause some additional stdout output due to the way we check if the
service still exists.

Therefore we pipe the error output to /dev/null.

=====

  Additional rpm output:
    Failed to get unit file state for obsapidelayed.service: No such file or directory

Co-authored-by: Eduardo Navarro <enavarro@suse.com>